PDF: Fix parsing of escaped name-object characters
Name object names can contain characters escaped by `#` signs, where
the `#` is followed by two hexadecimal values denoting the escaped
character's binary representation.

Prior to this change, the code would shift the first hex. value outside
the range of the byte in which it belonged, and would then also attempt
to interpret the decoded value as being outside of the name object,
possibly delimiting the next PDF object in the file.