FIX: Trim strings in WAVE Note and Label chunks.
Those strings are NULL terminated, calling trim() removes the NULL byte.

It also removes all trailing whitespace, which may not be desired.
Something like `str.replaceFirst("\0$", "")` would be more precise.
Anyway, trim() is used in this commit to be consistent with the rest of
the WAVE module where trim() is used as well.