AUDIT: Correct and align status reporting with other handlers
Audit reports weren't correctly differentiating between the standard
levels of validity in the report summaries, instead grouping
"well-formed" and "not well-formed" together, as well as ignoring
files of undetermined status.

This change correctly counts all current status types, replacing old
"not found" and "not processed" tallies (which were not being counted
correctly for the former, and at all for the latter) with a more
accurate tally of "unknown" statuses.

The concise report summaries have been altered to accommodate the new
tallies and also include documentation on how to interpret their values.