GUI: Allow directories to be analyzed from File menu
Also stops indiscriminately filtering out files which begin with
periods (.) on the false assumption that they should always be hidden.
This brings the behaviour in line with the command-line interface.

We *should* probably ignore files marked as hidden by default, but
we would first need to provide an option to disable that behaviour,
and make sure the implementation is system aware.