Fix tests for upgraded EPUBCheck, add info messages to JHOVE report
Two changes to EPUBCheck were causing the tests to change:
The first changes the error messages returned for validation of a non-EPUB:
https://github.com/w3c/epubcheck/pull/1134
The solution was to change the message count.
The second downgrades the encrypted file message to INFO
https://github.com/w3c/epubcheck/pull/1136
The original implementation of the JHOVE EPUB module did not include INFO messages generated by EPUBCheck in the JHOVE report, but looking at the kinds of messages that might be listed as INFO messages I thought it might be useful to include them. The change to add INFO messages is reflected in this commit.