PATCH: Failing tests on GitHub Actions
Some time in September the JHOVE XML module tests started to fail.
This was caused by a 403 error when trying to load XSD schema used in the two
XML test files in this PR. The schema references in the test files load
external schema located at <http://www.loc.gov/standards/>. These are behind
a Cloudflare proxy which is the source of the 403 for requests from GitHub
Actions only. The tests run fine when run locally, locally using different
JDKs via Docker, on GitHub actions machines locally and on Travis.

The solution has been to copy the schema to an OPF hosted domain for test
purposes <https://schema.openpreservation.org/ois/xml/xsd/jhove/test/>. A
robust implementation of schema checking in JHOVE is the real solution.