Bugfix in TiffIFD.readXMP method
When trying to extract the encoding part out of a string like
"ENC=B,UTF-8" (returned from the XMPHandler class) the encoding starts
at offset 6, not 5. We want "UTF-8", not ",UTF-8".

Without this fix JHOVE would raise a TIFF-HUL-14 error whenever there is
XMP metadata in a TIFF file that is enclosed in a packet wrapper (a pair
of XML processing instructions) and the packet wrapper has an encoding
attribute, regardless of which encoding is actually specified there
(because Java doesn't recognize the encoding string because of the
leading comma).

With this fix, JHOVE *should* raise a TIFF-HUL-14 error whenever there
is XMP metadata in a TIFF file that is enclosed by a packet wrapper, the
packet wrapper has an encoding attribute, and the encoding specified
that is not supported by Java. If it wasn't for another bug that was
uncovered by this bugfix, but we're getting there ...